fix tests in python2
