+++ src/bitmessageqt/networkstatus.py
@@ -15,7 +15,7 @@ import network.stats
 import shared
 from . import widgets
 from inventory import Inventory
-from network.connectionpool import BMConnectionPool
+from network.fix_circular_imports import BMConnectionPool
 from .retranslateui import RetranslateMixin
 from tr import _translate
 from .uisignaler import UISignaler
